src/poset/{pos,kvdb,trie,state} are moved to src/
