+++ src/kvdb/interface.go
