+++ src/kvdb/table_batch.go
