+++ src/pos/config.go
