+++ src/poset/store_debug.go
@@ -5,7 +5,7 @@ package poset
 import (
 	"github.com/Fantom-foundation/go-lachesis/src/common"
 	"github.com/Fantom-foundation/go-lachesis/src/peers"
-	"github.com/Fantom-foundation/go-lachesis/src/poset/state"
+	"github.com/Fantom-foundation/go-lachesis/src/state"
 )
 
 // Store provides an interface for persistent and non-persistent stores
