+++ src/state/journal.go
