+++ src/state/state_object.go
