+++ src/trie/encoding.go
