+++ src/trie/encoding_test.go
