+++ src/trie/hasher.go
