+++ src/trie/iterator.go
