+++ src/trie/node.go
