+++ src/trie/node_test.go
